:truck: Move SignalDeclarationStore class
