+++ Assets/Scripts/Implement/SignalHandler.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using JetBrains.Annotations;
 using UniRx;
 using Zenject;
@@ -40,26 +39,6 @@ namespace SignalHandler
                 SignalBus.Unsubscribe<TSignal>(OnReceived);
                 Subject.OnCompleted();
             }
-        }
-    }
-
-    internal static class SignalDeclarationStore
-    {
-        private static IDictionary<DiContainer, IList<Type>> DeclarationMap { get; } = new Dictionary<DiContainer, IList<Type>>();
-
-        internal static bool HasDeclaration<TSignal>(DiContainer container)
-        {
-            return DeclarationMap.ContainsKey(container) && DeclarationMap[container].Contains(typeof(TSignal));
-        }
-
-        internal static void AddDeclaration<TSignal>(DiContainer container)
-        {
-            if (!DeclarationMap.ContainsKey(container))
-            {
-                DeclarationMap[container] = new List<Type>();
-            }
-
-            DeclarationMap[container].Add(typeof(TSignal));
         }
     }
 }
